refactor: rename folder and add hover on list item
