+++ frontend/components/mention/MentionViewItem.tsx
@@ -1,7 +1,8 @@
 // SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2022 dv4all
+// SPDX-FileCopyrightText: 2023 - 2024 Netherlands eScience Center
 // SPDX-FileCopyrightText: 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
-// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2024 Dusan Mijatovic (Netherlands eScience Center)
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -54,7 +55,13 @@ export default function MentionViewItem({item, pos}: {item: MentionItemProps, po
 
   if (item?.url) {
     return (
-      <Link href={item.url} passHref target="_blank" rel="noreferrer">
+      <Link
+        href={item.url}
+        className="flex-1 pr-8"
+        target="_blank"
+        rel="noreferrer"
+        passHref
+      >
         {renderItemBody()}
       </Link>
     )
