+++ frontend/pages/admin/mentions.tsx
@@ -1,6 +1,6 @@
+// SPDX-FileCopyrightText: 2023 - 2024 Dusan Mijatovic (Netherlands eScience Center)
 // SPDX-FileCopyrightText: 2023 - 2024 Netherlands eScience Center
 // SPDX-FileCopyrightText: 2023 Christian Meeßen (GFZ) <christian.meessen@gfz-potsdam.de>
-// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
 // SPDX-FileCopyrightText: 2023 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2023 Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences
 // SPDX-FileCopyrightText: 2023 dv4all
@@ -14,7 +14,7 @@ import {app} from '../../config/app'
 import DefaultLayout from '~/components/layout/DefaultLayout'
 import AdminPageWithNav from '~/components/admin/AdminPageWithNav'
 import {adminPages} from '~/components/admin/AdminNav'
-import MentionsOverview from '~/components/admin/mentions-overview/MentionsOverview'
+import MentionsOverview from '~/components/admin/mentions/MentionsOverview'
 import {SearchProvider} from '~/components/search/SearchContext'
 import {PaginationProvider} from '~/components/pagination/PaginationContext'
